clearing post-user relation on deleting post

+++ controllers/feed.js
@@ -159,6 +159,13 @@ exports.updatePost = (req, res, next) => {
         throw err;
       }
 
+      // check post is created by current user?
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -199,12 +206,31 @@ exports.deletePost = (req, res, next) => {
         throw err;
       }
 
+      // check post is created by current user?
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
 
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      if (!user) {
+        const error = new Error("No user found.");
+        error.statusCode = 404;
+        throw error;
+      }
+
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
